refactor: private object 형태로 dimens 수정
