+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/calendar/CalendarScreen.kt
@@ -89,7 +89,7 @@ import org.moa.moa.presentation.ui.theme.WHITE
 import org.moa.moa.util.emotionRes
 import org.moa.moa.util.formatDateTime
 
-object CalendarDimens {
+private object CalendarDimens {
     const val TOTAL_DAY_CELLS = 42
     const val CALENDAR_FRACTION = 0.8f
     val verticalPadding = 40.dp
@@ -98,11 +98,7 @@ object CalendarDimens {
 
     const val BOTTOM_SHEET_CONTENT_HEIGHT = 0.5f
     val sheetVerticalPadding = 20.dp
-    val sheetDragHandleWidth = 68.dp
-    val sheetDragHandleHeight = 3.dp
-    val sheetDragHandleTopPadding = 5.dp
     val sheetShadowElevation = 8.dp
-    val sheetDragHandleRoundedCornerShape = RoundedCornerShape(2.5.dp)
 }
 
 @Composable
