+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/detail/component/CalendarDetailContentBackground.kt
@@ -2,6 +2,7 @@ package org.moa.moa.presentation.calendar.detail.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.runtime.Composable
@@ -11,12 +12,16 @@ import androidx.compose.ui.unit.dp
 import moa.presentation.generated.resources.Res
 import moa.presentation.generated.resources.calendar_detail_background
 import org.jetbrains.compose.resources.painterResource
-import org.moa.moa.presentation.calendar.detail.CalendarDetailDimens.CalendarDetailRoundedCorner
+import org.moa.moa.presentation.calendar.detail.component.ContentBackgroundDimens.contentRoundedCornerShape
+
+private object ContentBackgroundDimens{
+    val contentRoundedCornerShape = RoundedCornerShape(25.dp)
+}
 
 @Composable
 fun CalendarDetailContentBackground(modifier: Modifier) {
     Card(
-        shape = CalendarDetailRoundedCorner,
+        shape = contentRoundedCornerShape,
         elevation = CardDefaults.cardElevation(3.dp)
     ) {
         Image(
