+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/detail/component/CalendarDetailContentImage.kt
@@ -18,10 +18,14 @@ import coil3.compose.AsyncImage
 import moa.presentation.generated.resources.Res
 import moa.presentation.generated.resources.top_logo
 import org.jetbrains.compose.resources.painterResource
-import org.moa.moa.presentation.calendar.detail.CalendarDetailDimens.CONTENT_IMAGE_HEIGHT_FRACTION
+import org.moa.moa.presentation.calendar.detail.component.ContentImageDimens.CONTENT_IMAGE_HEIGHT_FRACTION
 import org.moa.moa.presentation.ui.theme.CORNER_RADIUS
 import org.moa.moa.presentation.ui.theme.GRAY3
 import org.moa.moa.presentation.ui.theme.GRAY4
+
+private object ContentImageDimens {
+    const val CONTENT_IMAGE_HEIGHT_FRACTION = 0.4f
+}
 
 @Composable
 fun CalendarDetailContentImage(
