--- presentation/src/commonMain/kotlin/org/moa/moa/presentation/home/HomeDimens.kt
@@ -1,11 +1,0 @@
-package org.moa.moa.presentation.home
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.ui.unit.dp
-
-object HomeDimens {
-    val verticalPadding = 20.dp
-    val buttonRoundedCornerShape = RoundedCornerShape(58.dp)
-    val buttonWidth = 222.dp
-    val buttonHeight = 54.dp
-}
