--- presentation/src/commonMain/kotlin/org/moa/moa/presentation/record/RecordDimens.kt
@@ -1,22 +1,0 @@
-package org.moa.moa.presentation.record
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-object RecordDimens {
-    const val RECORD_MAX_LENGTH = 500
-    const val RECORD_MAX_LINES = 20
-    const val CAMERA_IMAGE_FRACTION = 0.3f
-    const val RECORD_TIME_DELAY = 500L
-    val recorderButton = 86.dp
-    val recorderStopButton = 60.dp
-    val recordGuideText = 22.sp
-    val topPadding = 30.dp
-    val bottomPadding = 50.dp
-    val cameraButton = 46.dp
-    val cameraButtonText = 17.sp
-    val imagePadding = 20.dp
-    val shadow = 4.dp
-    val roundedCornerShape = RoundedCornerShape(15.dp)
-}
